Prepare for breaking changes in test libraries
Apparently the lovely folks at MSTest have decided to make breaking changes in a patch release (1.1.2). This includes renaming their packages and removing types that were apparently obsolete, but never marked as such. So, i'm making some adjustments to the usage here so that we can update to the new version cleanly. I've also collapsed the partial classes.